requirements.txt
llama-index==0.9.27
llama-index-core==0.12.5
llama-index-llms-azure-openai==0.1.3
llama-index-readers-file==0.1.3
python-dotenv==1.0.0
pydantic==2.5.2
typing-extensions>=4.8.0
python-dateutil>=2.8.2
